+++ client/src/components/modules/Chat.css
@@ -1,0 +1,13 @@
+/** @define Chat */
+.Chat-container {
+  height: 100%;
+}
+
+.Chat-newContainer {
+  width: 100%;
+}
+
+.Chat-historyContainer {
+  flex-grow: 1;
+  overflow-y: auto;
+}
