+++ client/src/components/modules/SingleChat.css
@@ -1,0 +1,21 @@
+/** @define SingleChat */
+.SingleChat-container {
+  margin: 0 0 0.25rem;
+}
+
+.SingleChat-sender {
+  width: 106px;
+  margin-right: 0.25rem;
+  flex-shrink: 0;
+}
+
+.SingleChat-content {
+  padding: 0.4rem 0.75rem;
+  background: var(--medgrey);
+  border-radius: 1rem;
+}
+
+.SingleChat-mine .SingleChat-content {
+  background: var(--primary--dim);
+  color: var(--white);
+}
