+++ client/src/components/modules/SingleUser.css
@@ -1,0 +1,15 @@
+/** @define SingleUser */
+.SingleUser-container {
+  background: var(--white);
+  border-radius: 4px;
+  padding: 1rem 1.5rem;
+  margin-bottom: 0.5rem;
+}
+
+.SingleUser-container:hover {
+  background: var(--medgrey);
+}
+
+.SingleUser-container--active {
+  background: var(--grey);
+}
